Fixing warning on boolean and > in graph_algorithm.h.
git-svn-id: http://svn.seqan.de/seqan/trunk@7987 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/projects/library/seqan/graph_algorithms/graph_algorithm.h
@@ -1368,9 +1368,9 @@ transitive_closure(Graph<TSpec> const& g,
 	for (TSize k=0;k<len;++k) {
 		for(TSize i=0;i<len;++i) {
 			for(TSize j=0;j<len;++j) {
-				bool t_ij = getValue(closure, i*len+j) > 0;
-				bool t_ik = getValue(closure, i*len+k) > 0;
-				bool t_kj = getValue(closure, k*len+j) > 0;
+				bool t_ij = static_cast<int>(getValue(closure, i*len+j)) > 0;
+                bool t_ik = static_cast<int>(getValue(closure, i*len+k)) > 0;
+                bool t_kj = static_cast<int>(getValue(closure, k*len+j)) > 0;
 				assignValue(local, i*len+j, t_ij || (t_ik && t_kj));
 			}
 		}
